refactor: move most common methods to `AbstractOperationHandler`
and make it implement `ExportHandler`.

This can now be extended for other bo handlers regardless of their exported entity
