+++ zeebe/exporters/camunda-exporter/src/main/java/io/camunda/exporter/handlers/batchoperation/AbstractOperationHandler.java
@@ -7,25 +7,33 @@
  */
 package io.camunda.exporter.handlers.batchoperation;
 
-import io.camunda.webapps.schema.entities.operation.OperationEntity;
+import io.camunda.exporter.handlers.ExportHandler;
+import io.camunda.webapps.schema.entities.ExporterEntity;
+import io.camunda.webapps.schema.entities.operation.OperationType;
+import io.camunda.zeebe.exporter.common.cache.ExporterEntityCache;
+import io.camunda.zeebe.exporter.common.cache.batchoperation.CachedBatchOperationEntity;
+import io.camunda.zeebe.protocol.record.Record;
+import io.camunda.zeebe.protocol.record.RecordValue;
+import io.camunda.zeebe.util.VisibleForTesting;
 
-public class AbstractOperationHandler {
+public abstract class AbstractOperationHandler<T extends ExporterEntity<T>, R extends RecordValue>
+    implements ExportHandler<T, R> {
 
   protected static final String ID_PATTERN = "%s_%s";
   protected final String indexName;
+  protected final ExporterEntityCache<String, CachedBatchOperationEntity> batchOperationCache;
+  @VisibleForTesting final OperationType relevantOperationType;
 
-  public AbstractOperationHandler(final String indexName) {
+  public AbstractOperationHandler(
+      final String indexName,
+      final ExporterEntityCache<String, CachedBatchOperationEntity> batchOperationCache,
+      final OperationType relevantOperationType) {
     this.indexName = indexName;
+    this.batchOperationCache = batchOperationCache;
+    this.relevantOperationType = relevantOperationType;
   }
 
-  public Class<OperationEntity> getEntityType() {
-    return OperationEntity.class;
-  }
-
-  public OperationEntity createNewEntity(final String id) {
-    return new OperationEntity().setId(id);
-  }
-
+  @Override
   public String getIndexName() {
     return indexName;
   }
@@ -40,5 +48,30 @@ public class AbstractOperationHandler {
    */
   protected String generateId(final long batchOperationKey, final long itemKey) {
     return String.format(ID_PATTERN, batchOperationKey, itemKey);
+  }
+
+  /**
+   * Checks if the record is relevant for the batch operation type handled by this handler. The
+   * record is relevant, when the operationType of the overall batch operation is the same as the
+   * monitored batch operation type of this record.<br>
+   * <br>
+   * This needs to be checked, because the <code>Record.getBatchOperationReference()</code> is
+   * present for all follow-up records of the actual operation command and not just the direct
+   * response record. E.g. an incident present on a canceled process instance is also resolved
+   * during the cancellation and the appended <code>Incident.RESOLVED</code> also has a valid
+   * batchOperationReference.
+   *
+   * @param record the record to check for relevance
+   * @return true if the record is relevant, otherwise false
+   */
+  protected boolean isRelevantForBatchOperation(final Record<R> record) {
+    final var cachedEntity =
+        batchOperationCache.get(String.valueOf(record.getBatchOperationReference()));
+
+    return cachedEntity
+        .filter(
+            cachedBatchOperationEntity ->
+                cachedBatchOperationEntity.type() == relevantOperationType)
+        .isPresent();
   }
 }
