+++ zeebe/exporters/camunda-exporter/src/main/java/io/camunda/exporter/handlers/batchoperation/AbstractOperationStatusHandler.java
@@ -10,7 +10,6 @@ package io.camunda.exporter.handlers.batchoperation;
 import static io.camunda.zeebe.protocol.record.RecordMetadataDecoder.batchOperationReferenceNullValue;
 
 import io.camunda.exporter.exceptions.PersistenceException;
-import io.camunda.exporter.handlers.ExportHandler;
 import io.camunda.exporter.store.BatchRequest;
 import io.camunda.webapps.schema.descriptors.template.OperationTemplate;
 import io.camunda.webapps.schema.entities.operation.OperationEntity;
@@ -22,7 +21,6 @@ import io.camunda.zeebe.protocol.record.Record;
 import io.camunda.zeebe.protocol.record.RecordValue;
 import io.camunda.zeebe.protocol.record.ValueType;
 import io.camunda.zeebe.util.DateUtil;
-import io.camunda.zeebe.util.VisibleForTesting;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -38,30 +36,30 @@ import org.slf4j.LoggerFactory;
  * @param <R> the type of the record value handled by this operation status handler
  */
 public abstract class AbstractOperationStatusHandler<R extends RecordValue>
-    extends AbstractOperationHandler implements ExportHandler<OperationEntity, R> {
+    extends AbstractOperationHandler<OperationEntity, R> {
 
   public static final String ERROR_MSG = "%s: %s";
   private static final Logger LOGGER =
       LoggerFactory.getLogger(AbstractOperationStatusHandler.class);
   protected final ValueType handledValueType;
-  protected final ExporterEntityCache<String, CachedBatchOperationEntity> batchOperationCache;
-  @VisibleForTesting final OperationType relevantOperationType;
 
   public AbstractOperationStatusHandler(
       final String indexName,
       final ValueType handledValueType,
       final OperationType relevantOperationType,
       final ExporterEntityCache<String, CachedBatchOperationEntity> batchOperationCache) {
-    super(indexName);
+    super(indexName, batchOperationCache, relevantOperationType);
     this.handledValueType = handledValueType;
-    this.relevantOperationType = relevantOperationType;
-
-    this.batchOperationCache = batchOperationCache;
   }
 
   @Override
   public ValueType getHandledValueType() {
     return handledValueType;
+  }
+
+  @Override
+  public Class<OperationEntity> getEntityType() {
+    return OperationEntity.class;
   }
 
   @Override
@@ -74,6 +72,11 @@ public abstract class AbstractOperationStatusHandler<R extends RecordValue>
   @Override
   public List<String> generateIds(final Record<R> record) {
     return List.of(generateId(record.getBatchOperationReference(), getItemKey(record)));
+  }
+
+  @Override
+  public OperationEntity createNewEntity(final String id) {
+    return new OperationEntity().setId(id);
   }
 
   @Override
@@ -108,11 +111,6 @@ public abstract class AbstractOperationStatusHandler<R extends RecordValue>
     LOGGER.trace("Updated operation {} with fields {}", entity.getId(), updateFields);
   }
 
-  @Override
-  public String getIndexName() {
-    return indexName;
-  }
-
   /**
    * Extract the operation itemKey from the record.
    *
@@ -134,29 +132,4 @@ public abstract class AbstractOperationStatusHandler<R extends RecordValue>
 
   /** Checks if the batch operation item is failed */
   abstract boolean isFailed(Record<R> record);
-
-  /**
-   * Checks if the record is relevant for the batch operation type handled by this handler. The
-   * record is relevant, when the operationType of the overall batch operation is the same as the
-   * monitored batch operation type of this record.<br>
-   * <br>
-   * This needs to be checked, because the <code>Record.getBatchOperationReference()</code> is
-   * present for all follow-up records of the actual operation command and not just the direct
-   * response record. E.g. an incident present on a canceled process instance is also resolved
-   * during the cancellation and the appended <code>Incident.RESOLVED</code> also has a valid
-   * batchOperationReference.
-   *
-   * @param record the record to check for relevance
-   * @return true if the record is relevant, otherwise false
-   */
-  boolean isRelevantForBatchOperation(final Record<R> record) {
-    final var cachedEntity =
-        batchOperationCache.get(String.valueOf(record.getBatchOperationReference()));
-
-    return cachedEntity
-        .filter(
-            cachedBatchOperationEntity ->
-                cachedBatchOperationEntity.type() == relevantOperationType)
-        .isPresent();
-  }
 }
